+++ NEWS.md
@@ -1,3 +1,8 @@
+# kim 0.5.204
+* Added the following function(s): contingency_table, chi_squared_test
+* Updated the following function(s) (and other functions that use them): 
+chi_squared_test_pairwise
+
 # kim 0.5.202
 * Critical fix or improvement in the following functions (and other 
 functions that use them): z_score*